Added RestController and services

+++ src/main/java/com/esprit/spring/etude_de_cas_spring/repositories/AbonnementRepository.java
@@ -2,5 +2,5 @@ package com.esprit.spring.etude_de_cas_spring.repositories;
 import com.esprit.spring.etude_de_cas_spring.entities.Abonnement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AbonnementRepository extends JpaRepository<Abonnement,Long> {
+public interface AbonnementRepository extends JpaRepository<Abonnement,Integer> {
 }
